+++ utils/maybe/maybe.go
@@ -65,12 +65,5 @@ func Bind[T, U any](m Maybe[T], f func(T) U) Maybe[U] {
 
 // Equal returns true if both m1 and m2 are nothing or have the same value according to [equalFunc].
 func Equal[T any](m1 Maybe[T], m2 Maybe[T], equalFunc func(T, T) bool) bool {
-	if m1.IsNothing() {
-		return m2.IsNothing()
-	}
-
-	if m2.IsNothing() {
-		return false
-	}
-	return equalFunc(m1.Value(), m2.Value())
+	return m1.IsNothing() == m2.IsNothing() && equalFunc(m1.Value(), m2.Value())
 }
