+++ x/merkledb/proof.go
@@ -70,22 +70,16 @@ func (node *ProofNode) ToProto() *pb.ProofNode {
 			NibbleLength: uint64(node.KeyPath.NibbleLength),
 			Value:        node.KeyPath.Value,
 		},
+		ValueOrHash: &pb.MaybeBytes{
+			Value:     node.ValueOrHash.Value(),
+			IsNothing: node.ValueOrHash.IsNothing(),
+		},
 		Children: make(map[uint32][]byte, len(node.Children)),
 	}
 
 	for childIndex, childID := range node.Children {
 		childID := childID
 		pbNode.Children[uint32(childIndex)] = childID[:]
-	}
-
-	if node.ValueOrHash.HasValue() {
-		pbNode.ValueOrHash = &pb.MaybeBytes{
-			Value: node.ValueOrHash.Value(),
-		}
-	} else {
-		pbNode.ValueOrHash = &pb.MaybeBytes{
-			IsNothing: true,
-		}
 	}
 
 	return pbNode
@@ -568,21 +562,12 @@ func (proof *ChangeProof) ToProto() *pb.ChangeProof {
 
 	keyChanges := make([]*pb.KeyChange, len(proof.KeyChanges))
 	for i, kv := range proof.KeyChanges {
-		var value pb.MaybeBytes
-		if kv.Value.HasValue() {
-			value = pb.MaybeBytes{
-				Value:     kv.Value.Value(),
-				IsNothing: false,
-			}
-		} else {
-			value = pb.MaybeBytes{
-				Value:     nil,
-				IsNothing: true,
-			}
-		}
 		keyChanges[i] = &pb.KeyChange{
-			Key:   kv.Key,
-			Value: &value,
+			Key: kv.Key,
+			Value: &pb.MaybeBytes{
+				Value:     kv.Value.Value(),
+				IsNothing: kv.Value.IsNothing(),
+			},
 		}
 	}
