+++ x/merkledb/proof_test.go
@@ -1373,11 +1373,11 @@ func TestProofNodeProtoMarshalUnmarshal(t *testing.T) {
 		protoNode := node.ToProto()
 		var unmarshaledNode ProofNode
 		require.NoError(unmarshaledNode.UnmarshalProto(protoNode))
-		require.Equal(node, unmarshaledNode)
+		require.True(proofNodeEqual(node, unmarshaledNode))
 
 		// Marshaling again should yield same result.
 		protoUnmarshaledNode := unmarshaledNode.ToProto()
-		require.Equal(protoNode, protoUnmarshaledNode)
+		require.True(protoProofNodeEqual(protoNode, protoUnmarshaledNode))
 	}
 }
 
@@ -1498,216 +1498,6 @@ func TestChangeProofProtoMarshalUnmarshal(t *testing.T) {
 	}
 }
 
-func protoRangeProofsEqual(proof1, proof2 *pb.RangeProof) bool {
-	if (proof1 == nil && proof2 != nil) || (proof2 == nil && proof1 != nil) {
-		return false
-	}
-	if proof1 == nil && proof2 == nil {
-		return true
-	}
-	if len(proof1.Start) != len(proof2.Start) {
-		return false
-	}
-	if len(proof1.End) != len(proof2.End) {
-		return false
-	}
-	if len(proof1.KeyValues) != len(proof2.KeyValues) {
-		return false
-	}
-
-	for i := 0; i < len(proof1.Start); i++ {
-		if !protoProofNodeEqual(proof1.Start[i], proof2.Start[i]) {
-			return false
-		}
-	}
-	for i := 0; i < len(proof1.End); i++ {
-		if !protoProofNodeEqual(proof1.End[i], proof2.End[i]) {
-			return false
-		}
-	}
-	for i := 0; i < len(proof1.KeyValues); i++ {
-		if !protoKeyValueEqual(proof1.KeyValues[i], proof2.KeyValues[i]) {
-			return false
-		}
-	}
-	return true
-}
-
-func protoChangeProofsEqual(proof1, proof2 *pb.ChangeProof) bool {
-	if (proof1 == nil && proof2 != nil) || (proof2 == nil && proof1 != nil) {
-		return false
-	}
-	if proof1 == nil && proof2 == nil {
-		return true
-	}
-	if len(proof1.StartProof) != len(proof2.StartProof) {
-		return false
-	}
-	if len(proof1.EndProof) != len(proof2.EndProof) {
-		return false
-	}
-	if len(proof1.KeyChanges) != len(proof2.KeyChanges) {
-		return false
-	}
-
-	for i := 0; i < len(proof1.StartProof); i++ {
-		if !protoProofNodeEqual(proof1.StartProof[i], proof2.StartProof[i]) {
-			return false
-		}
-	}
-	for i := 0; i < len(proof1.EndProof); i++ {
-		if !protoProofNodeEqual(proof1.EndProof[i], proof2.EndProof[i]) {
-			return false
-		}
-	}
-	for i := 0; i < len(proof1.KeyChanges); i++ {
-		if !protoKeyChangeEqual(proof1.KeyChanges[i], proof2.KeyChanges[i]) {
-			return false
-		}
-	}
-	return true
-}
-
-func protoProofNodeEqual(node1, node2 *pb.ProofNode) bool {
-	if (node1 == nil && node2 != nil) || (node2 == nil && node1 != nil) {
-		return false
-	}
-	if node1 == nil && node2 == nil {
-		return true
-	}
-
-	return protoSerializedPathEqual(node1.Key, node2.Key) &&
-		protoMaybeBytesEqual(node1.ValueOrHash, node2.ValueOrHash) &&
-		protoChildrenEqual(node1.Children, node2.Children)
-}
-
-func protoSerializedPathEqual(path1, path2 *pb.SerializedPath) bool {
-	return (path1 == nil && path2 == nil) || (path1 != nil && path2 != nil && bytes.Equal(path1.Value, path2.Value) && path1.NibbleLength == path2.NibbleLength)
-}
-
-func protoKeyValueEqual(change1, change2 *pb.KeyValue) bool {
-	return (change1 == nil && change2 == nil) || (change1 != nil && change2 != nil && bytes.Equal(change1.Key, change2.Key) && bytes.Equal(change1.Value, change2.Value))
-}
-
-func protoKeyChangeEqual(change1, change2 *pb.KeyChange) bool {
-	return (change1 == nil && change2 == nil) || (change1 != nil && change2 != nil && bytes.Equal(change1.Key, change2.Key) && protoMaybeBytesEqual(change1.Value, change2.Value))
-}
-
-func protoMaybeBytesEqual(maybe1, maybe2 *pb.MaybeBytes) bool {
-	if (maybe1 == nil && maybe2 != nil) || (maybe2 == nil && maybe1 != nil) {
-		return false
-	}
-	if maybe1 == nil && maybe2 == nil {
-		return true
-	}
-
-	return maybe1.IsNothing == maybe2.IsNothing && bytes.Equal(maybe1.Value, maybe2.Value)
-}
-
-func protoChildrenEqual(children1, children2 map[uint32][]byte) bool {
-	if len(children1) != len(children2) {
-		return false
-	}
-	for key, value := range children1 {
-		if otherVal, ok := children2[key]; !ok || !bytes.Equal(otherVal, value) {
-			return false
-		}
-	}
-	return true
-}
-
-func rangeProofsEqual(proof1, proof2 RangeProof) bool {
-	if len(proof1.StartProof) != len(proof2.StartProof) {
-		return false
-	}
-	if len(proof1.EndProof) != len(proof2.EndProof) {
-		return false
-	}
-	if len(proof1.KeyValues) != len(proof2.KeyValues) {
-		return false
-	}
-
-	for i := 0; i < len(proof1.StartProof); i++ {
-		if !proofNodeEqual(proof1.StartProof[i], proof2.StartProof[i]) {
-			return false
-		}
-	}
-	for i := 0; i < len(proof1.EndProof); i++ {
-		if !proofNodeEqual(proof1.EndProof[i], proof2.EndProof[i]) {
-			return false
-		}
-	}
-	for i := 0; i < len(proof1.KeyValues); i++ {
-		if !keyValueEqual(proof1.KeyValues[i], proof2.KeyValues[i]) {
-			return false
-		}
-	}
-	return true
-}
-
-func changeProofsEqual(proof1, proof2 ChangeProof) bool {
-	if len(proof1.StartProof) != len(proof2.StartProof) {
-		return false
-	}
-	if len(proof1.EndProof) != len(proof2.EndProof) {
-		return false
-	}
-	if len(proof1.KeyChanges) != len(proof2.KeyChanges) {
-		return false
-	}
-
-	for i := 0; i < len(proof1.StartProof); i++ {
-		if !proofNodeEqual(proof1.StartProof[i], proof2.StartProof[i]) {
-			return false
-		}
-	}
-	for i := 0; i < len(proof1.EndProof); i++ {
-		if !proofNodeEqual(proof1.EndProof[i], proof2.EndProof[i]) {
-			return false
-		}
-	}
-	for i := 0; i < len(proof1.KeyChanges); i++ {
-		if !keyChangeEqual(proof1.KeyChanges[i], proof2.KeyChanges[i]) {
-			return false
-		}
-	}
-	return true
-}
-
-func proofNodeEqual(node1, node2 ProofNode) bool {
-	return serializedPathEqual(node1.KeyPath, node2.KeyPath) &&
-		maybeBytesEqual(node1.ValueOrHash, node2.ValueOrHash) &&
-		childrenEqual(node1.Children, node2.Children)
-}
-
-func serializedPathEqual(path1, path2 SerializedPath) bool {
-	return bytes.Equal(path1.Value, path2.Value) && path1.NibbleLength == path2.NibbleLength
-}
-
-func keyValueEqual(change1, change2 KeyValue) bool {
-	return bytes.Equal(change1.Key, change2.Key) && bytes.Equal(change1.Value, change2.Value)
-}
-
-func keyChangeEqual(change1, change2 KeyChange) bool {
-	return bytes.Equal(change1.Key, change2.Key) && maybeBytesEqual(change1.Value, change2.Value)
-}
-
-func maybeBytesEqual(maybe1, maybe2 maybe.Maybe[[]byte]) bool {
-	return maybe1.IsNothing() == maybe2.IsNothing() && bytes.Equal(maybe1.Value(), maybe2.Value())
-}
-
-func childrenEqual(children1, children2 map[byte]ids.ID) bool {
-	if len(children1) != len(children2) {
-		return false
-	}
-	for key, value := range children1 {
-		if otherVal, ok := children2[key]; !ok || !bytes.Equal(otherVal[:], value[:]) {
-			return false
-		}
-	}
-	return true
-}
-
 func TestChangeProofUnmarshalProtoNil(t *testing.T) {
 	var proof ChangeProof
 	err := proof.UnmarshalProto(nil)
@@ -1824,11 +1614,11 @@ func TestProofProtoMarshalUnmarshal(t *testing.T) {
 		var unmarshaledProof Proof
 		protoProof := proof.ToProto()
 		require.NoError(unmarshaledProof.UnmarshalProto(protoProof))
-		require.Equal(proof, unmarshaledProof)
+		require.True(proofsEqual(proof, unmarshaledProof))
 
 		// Marshaling again should yield same result.
 		protoUnmarshaledProof := unmarshaledProof.ToProto()
-		require.Equal(protoProof, protoUnmarshaledProof)
+		require.True(protoProofsEqual(protoProof, protoUnmarshaledProof))
 	}
 }
 
@@ -1872,43 +1662,40 @@ func TestProofProtoUnmarshal(t *testing.T) {
 }
 
 func FuzzRangeProofInvariants(f *testing.F) {
-	now := time.Now().UnixNano()
-	f.Logf("seed: %d", now)
-	rand := rand.New(rand.NewSource(now)) // #nosec G404
-
-	var (
-		numKeyValues  = 2048
-		deletePortion = 0.25
-	)
-
-	db, err := getBasicDB()
-	require.NoError(f, err)
-
-	// Insert a bunch of random key values.
-	insertRandomKeyValues(
-		require.New(f),
-		rand,
-		[]database.Database{db},
-		numKeyValues,
-		deletePortion,
-	)
-
+	var deletePortion = 0.25
 	f.Fuzz(func(
 		t *testing.T,
+		randSeed int64,
 		startBytes []byte,
 		endBytes []byte,
 		maxProofLen uint,
+		numKeyValues uint,
 	) {
 		require := require.New(t)
 
-		// Make sure proof bounds are valid
-		if len(endBytes) != 0 && bytes.Compare(startBytes, endBytes) > 0 {
-			return
-		}
 		// Make sure proof length is valid
 		if maxProofLen == 0 {
-			return
+			t.SkipNow()
 		}
+
+		// Make sure proof bounds are valid
+		if len(endBytes) != 0 && bytes.Compare(startBytes, endBytes) > 0 {
+			t.SkipNow()
+		}
+
+		rand := rand.New(rand.NewSource(randSeed)) // #nosec G404
+
+		db, err := getBasicDB()
+		require.NoError(err)
+
+		// Insert a bunch of random key values.
+		insertRandomKeyValues(
+			require,
+			rand,
+			[]database.Database{db},
+			int(numKeyValues),
+			deletePortion,
+		)
 
 		start := maybe.Nothing[[]byte]()
 		if len(startBytes) != 0 {
@@ -1957,7 +1744,7 @@ func FuzzRangeProofInvariants(f *testing.F) {
 			if len(rangeProof.KeyValues) == 0 {
 				if len(rangeProof.StartProof) == 0 {
 					require.Len(rangeProof.EndProof, 1) // Just the root
-					require.Empty(rangeProof.EndProof[0].KeyPath)
+					require.Empty(rangeProof.EndProof[0].KeyPath.Value)
 				} else {
 					require.Empty(rangeProof.EndProof)
 				}
@@ -2004,32 +1791,28 @@ func FuzzRangeProofInvariants(f *testing.F) {
 }
 
 func FuzzProof(f *testing.F) {
-	now := time.Now().UnixNano()
-	f.Logf("seed: %d", now)
-	rand := rand.New(rand.NewSource(now)) // #nosec G404
-
 	var (
-		numKeyValues  = 2048
 		deletePortion = 0.25
 	)
-
-	db, err := getBasicDB()
-	require.NoError(f, err)
-
-	// Insert a bunch of random key values.
-	insertRandomKeyValues(
-		require.New(f),
-		rand,
-		[]database.Database{db},
-		numKeyValues,
-		deletePortion,
-	)
-
 	f.Fuzz(func(
 		t *testing.T,
 		key []byte,
+		randSee int64,
+		numKeyValues uint,
 	) {
+		rand := rand.New(rand.NewSource(randSee)) // #nosec G404
 		require := require.New(t)
+		db, err := getBasicDB()
+		require.NoError(err)
+
+		// Insert a bunch of random key values.
+		insertRandomKeyValues(
+			require,
+			rand,
+			[]database.Database{db},
+			int(numKeyValues),
+			deletePortion,
+		)
 
 		proof, err := db.GetProof(
 			context.Background(),
@@ -2060,41 +1843,40 @@ func FuzzProof(f *testing.F) {
 
 // Generate change proofs and verify that they are valid.
 func FuzzChangeProof(f *testing.F) {
-	now := time.Now().UnixNano()
-	f.Logf("seed: %d", now)
-	rand := rand.New(rand.NewSource(now)) // #nosec G404
 
 	const (
 		numKeyValues  = defaultHistoryLength / 2
 		deletePortion = 0.25
 	)
 
-	db, err := getBasicDB()
-	require.NoError(f, err)
-
-	startRootID, err := db.GetMerkleRoot(context.Background())
-	require.NoError(f, err)
-
-	// Insert a bunch of random key values.
-	// Don't insert so many that we have insufficient history.
-	insertRandomKeyValues(
-		require.New(f),
-		rand,
-		[]database.Database{db},
-		numKeyValues,
-		deletePortion,
-	)
-
-	endRootID, err := db.GetMerkleRoot(context.Background())
-	require.NoError(f, err)
-
 	f.Fuzz(func(
 		t *testing.T,
 		startBytes []byte,
 		endBytes []byte,
 		maxProofLen uint,
+		randSeed int64,
 	) {
 		require := require.New(t)
+		rand := rand.New(rand.NewSource(randSeed)) // #nosec G404
+
+		db, err := getBasicDB()
+		require.NoError(err)
+
+		startRootID, err := db.GetMerkleRoot(context.Background())
+		require.NoError(err)
+
+		// Insert a bunch of random key values.
+		// Don't insert so many that we have insufficient history.
+		insertRandomKeyValues(
+			require,
+			rand,
+			[]database.Database{db},
+			numKeyValues,
+			deletePortion,
+		)
+
+		endRootID, err := db.GetMerkleRoot(context.Background())
+		require.NoError(err)
 
 		// Make sure proof bounds are valid
 		if len(endBytes) != 0 && bytes.Compare(startBytes, endBytes) > 0 {
